# frontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PolarDrive.WebApi",
    "version": "v1"
  },
  "paths": {
...
    "/api/ClientCompanies": {
      "get": {
        "tags": ["ClientCompanies"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientCompanyDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientCompanyDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientCompanyDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["ClientCompanies"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCompanyDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCompanyDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCompanyDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ClientCompanies/{id}": {
      "put": {
        "tags": ["ClientCompanies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCompanyDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCompanyDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCompanyDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/ClientVehicles": {
      "get": {
        "tags": ["ClientVehicles"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminWorkflowExtendedDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminWorkflowExtendedDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminWorkflowExtendedDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["ClientVehicles"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientVehicleDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientVehicleDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientVehicleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ClientVehicles/{id}/status": {
      "patch": {
        "tags": ["ClientVehicles"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleStatusUpdateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleStatusUpdateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleStatusUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ClientVehicles/{id}": {
      "put": {
        "tags": ["ClientVehicles"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientVehicleDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientVehicleDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientVehicleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminWorkflowExtendedDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "vin": {
            "type": "string",
            "nullable": true
          },
          "fuelType": {
            "type": "string",
            "nullable": true
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "trim": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "clientOAuthAuthorized": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "isFetching": {
            "type": "boolean"
          },
          "firstActivationAt": {
            "type": "string",
            "nullable": true
          },
          "lastDeactivationAt": {
            "type": "string",
            "nullable": true
          },
          "lastFetchingDataAt": {
            "type": "string",
            "nullable": true
          },
          "clientCompany": {
            "$ref": "#/components/schemas/ClientCompanyDTO"
          }
        },
        "additionalProperties": false
      },
      "ClientCompanyDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "vatNumber": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "pecAddress": {
            "type": "string",
            "nullable": true
          },
          "landlineNumber": {
            "type": "string",
            "nullable": true
          },
          "referentName": {
            "type": "string",
            "nullable": true
          },
          "vehicleMobileNumber": {
            "type": "string",
            "nullable": true
          },
          "referentEmail": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ClientVehicleDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "clientCompanyId": {
            "type": "integer",
            "format": "int32"
          },
          "vin": {
            "type": "string",
            "nullable": true
          },
          "fuelType": {
            "type": "string",
            "nullable": true
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "trim": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "isFetching": {
            "type": "boolean"
          },
          "firstActivationAt": {
            "type": "string",
            "nullable": true
          },
          "lastDeactivationAt": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "VehicleStatusUpdateDTO": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          },
          "isFetching": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}